Refactored static methods from CoverageData into a new class CoverageUtils
